Base classes for passing encryption information to `HiveSplit`
Introducing basic constructs for fetching and passing encryption information to HiveSplit.

`HiveEncryptionInformationProvider` loops over available `EncryptionInformationSource` to
wait for the first one to return `EncryptionInformation`. `EncryptionInformation` contains
member variables which contains information about the different encryption formats.
Currently only `DWRF` encryption is supported and hence it is the only member variable in it.