Fix flaky TestFileSingleStreamSpiller
The test was sporadically failing at

> assertEquals(listFiles(spillPath.toPath()).size(), 0);

The assertion verifies that the spiller does the cleanup on close.

Since the test is not single threaded, test methods are allowed to
run in parallel. Thus it is not correct to write to a single shared
folder. Instead each test method has to write to it's own folder to
guarantee that the folder will be empty after the cleanup.