Fix integer overflow error in Bing Tiles
Since Java only has signed ints, zoom 23 was causing int overflows
in certain cases. In particular, if you are look at pixels at the
eastern/southern edges of the world, 256 << 23 == 2 ** 31, while
max int is 2 ** 32 - 1.

We fix this by doing these calculations in tiles, for which the
corresponding number of 1 << 23, which is within an int's bounds.