Fix bytecode generation for SQL functions
A lambda expression inside the body of a SQL function might have
the same argument name, as one of the arguments of the function
itself. Bytecode generation fails in this case.

Example:
CREATE FUNCTION testing.test.array_sum(x array<int>) RETURNS int
RETURN reduce(x, 0, (s, x) -> s + x, s -> s)

Here "x" is an argument for both the function and the lambda
expression in the function body. This change fixes the bug by
properly scoping variables in the lambda expression body, so that
there is no conflict with variables of the same name defined
outside the lambda expression's scope.