Fix infinite loop in scheduler for finished query
Queries that don't have table scan inputs could enter an infinite loop in
the scheduler if they hit the queryStateMachine.isDone() check and had any
sections ready for execution. If the query is done (e.g. canceled or
abandoned at this point), we abort the section.  Since "aborted" is
a failed state, the section would continue to be considered "ready for
execution"  in our next time through, and the loop would continue indefitely.
We need to explicitly exit the scheduling loop if the query is finished.

Queries with inputs didn't have this issue because the split scheduler
would be closed when the query finished, so the scheduler would hit an
error when it tried to create the scheduler for the scan stages.