Add geometry_to_dissolved_bing_tiles
A useful feature (that is requested by users) is to have a tile covering
of a geometry that recursively dissolves complete sets of children into their
parents.  This results in a smaller set of tiles with non-uniform zoom
level.  This can be done efficiently, and actually can form the basis
of the current (uniform zoom level) minimal tile covering function.

This commit implements geometry_to_dissolved_bing_tiles.  Later commits
will refactor geometry_to_bing_tiles to use the intermediate state for a
significant efficiency win.