Add support in parquet reader for reading TIMESTAMP_MICROS type.
Summary:
Right now presto always assumes TIMESTAMP_MILLIS as the OriginalType when int64 is used to
represent timestamp type in the schema. This causes an issue.

When we use createParquetPageSource we use our own type convertor and this does not have a check
for TIMESTAMP_MICROS and this can fail the query

Fix: In this change I am adding three ValueDecoders one for plain and one for RLE compression format and
one for the non batch reader. All of these are in 3 new classes so that during creation we
instantiate the class based on the OriginalType. One Alternate approach is to do this check inside
LongValueDecoders but that would make this check inside the most critical path and can affect query
performance. The fix simply divides the micros seconds by 1000 to get the milliseconds and this should
be ok because in presto we anyway operate at the millisecond granularity. Note that the non batch reader
also checks for whether timezone is present in the data while the batch readers don't because that
information is not available.

For testing, the ValueDecoders have their own unit test. Apart from that have also added a new test in
ParquetTester to check if timestamp stored as int64 with OriginalType TIMESTAMP_MICROS works.

Note: We use hive parquet writer and schema definition as defined in the presto-hive package
which is a shaded jar with a old version of hive. This packs an old version of parquet-mr that does not
have the TIMESTAMP_MICROS type. However, in the read path, we use the independent parquet dependency
which is more recent and so we use the version where the new enum is available. If we upgrade all the tests
to use the new version of parquet, we would not be testing parquet version written
by the packaged hive writer. We would need a version of the test that would work with the new definitions.
For now I have done this only for the timestamp type to test TIMESTAMP_MICROS. As part of seperate change
we can make sure all the test are tested with latest parquet writer. This change uses ExampleParquetWriter
which is supposed to be used only for demo or test purposes. We cannot use presto's parquet writer as well because
the one used in test always stored timestamp as millis and does not store any annotated type and using
that we will not be able to test our code.