Add test infrastructure to run queries using native workers
Extend HiveQueryRunner and DistributedQueryRunner to allow for setting up a
Java-based coordinator and a set of native workers. Native workers are created
by launching separate processes using native binary. The path to the binary is
specified using PRESTO_SERVER environment variable. The new test that uses
this infrastructure, TestHiveNativeWorkersQueries.java, is disabled by
default. In the future, we'll figure out how to get native worker binary on a
CI machine to enable this test.

The new test passed locally.