Fix client-info test-name output
The serialization of the Optional test-name field was outputting an
object with its only key value pair being "Present". We want the
actual test-name string to be serialized to json instead. Airlift's
toJson seems to handle serialization a bit better than ObjectMapper's
however Airlift's toJson pretty prints the json with newlines. This
will unfortunately break if we pass the client info into a header
since newlines are not allowed, so these should be stripped.