Use more efficient getApproximateLogicalSizeInBytes in OrcWriter
OrcWriter.estimateAverageLogicalSizePerRow has the top memory allocation
in some KDS pipeline. Since the original code was to estimate the page's
logical size and do not require accurate size to be calculated, this
commit changes the call of the expensive getLogicalSizeInBytes method to
a much faster and momory friendly method getApproximateLogicalSizeInBytes
that estimate the approximate logical size of a page.