Dereference SerializedPageReferences in batches
Refactors SerializedPageReference and usage sites to perform dereferencing
work in batches which means only a single call into the synchronized memory
tracker update logic will occur. Incidentally, this reduces a per page lambda
creation since the callback to invoke when the reference count hits 0 is
now created on a per output buffer basis, and created the opportunity to share
a single implementation for tracking lifespans and callbacks.