Fix output counting in OptimizedPartitionedOutputOperator
OptimizedPartitionedOutputOperator used to report the output data size
as the input data size. However, output data size might be larger due
to loss of run-length and/or dictionary encodings. Furthermore, the
output position count for the replicated case used to be the same as
the input position count, while the correct value should be the input
position count times partition count. This commit fixes this issue by
reporting the output sizes right before flushing each partition when
its buffer is full, where the accurate data size and positionCount can
be obtained.