Refactor DBResourceGroupConfigurationManager to support different data providers
We want a manager that can reload resource group configuration from data stored in various places, but the current DbResourceGroupsConfigurationManager allows only for data retrieval from a database. So, we are refactoring this class so that we can reuse the code that maintains the resource groups configurations, updating and disabling them as needed, and allow for different data providers to read the basic configurations.

The following decisions were made while thinking about the refactored design:

DbResourceGroupConfigurationManager refactored to ReloadingResourceGroupConfigurationManager, which accepts a ManagerSpecProvider to be passed in that provides the basic resource group configuration definitions to the manager.
For the DB based manager, DbManagerSpecProvider handles reading from the database and provides the basic spec to the ReloadingResourceGroupConfigurationManager