Improve SOFT_AFFINITY strategy to be more resilient to node downtime
Currently with SOFT_AFFINITY, if a worker goes down the
cache is re-distributed among the alive workers. This will
affect all workers and cache hit rate will be bad. This is bad
especially for the cases where the worker comes back up in 10-15min.
With this PR, given a split, we try to find a worker from
a cached list of workers, and if that worker is down, we
schedule the split to another worker. So cache hit rate
is bad only for that one died worker. Rest all workers
remain untouched.