Support enforcement of NOT NULL column declarations
This PR is back ported from Trino
https://github.com/trinodb/trino/pull/4144, originally authoered by
djsstarburst. Quote description from Trino PR4144
"This commit enforces NOT NULL column declarations on write
in the Presto engine, so it applies to all connectors. The
existing Postgres and Mysql tests named testInsertIntoNotNullColumn
were changed to check for the new error message, and a new test
with the same name was added to TestIcebergSmoke.

One possible concern with this commit is that the error message
issued by the Presto engine when writing a null to a NOT NULL
column is a different message than the Connector might issue
if no value was supplied for the NOT NULL column. I think this
is ok, because the error messages supplied by the Connectors are
completely specific to the Connector."

Because the gap between Presto and Trino, the original PR has to be
modified.

Cherry-pick of https://github.com/trinodb/trino/pull/4144 (
https://github.com/trinodb/trino/pull/4144)

Co-authored-by: djsstarburst <djsstarburst@users.noreply.github.com>