Update /v1/info/state API
Update the API to return ACTIVE when the active worker count is greater or equals to the minimum worker count for Coordinator
No change in behavior for worker

Tested with:
1) No cluster.required-workers-active
2) cluster.required-workers-active set to 0
3) cluster.required-workers-active set to a positive number

In all three cases, the cluster was coming up and running properly

Also tested the usecase when the active workers drop below the threshold and confirmed that the state returns properly.

Checked the EP on the worker node as well.