Do not create local projection for constant in PlanRemoteProjections
Previously in PlanRemoteProjection, we might create extra assignments for
constants. This would cause plan failure in case of dereference. Essentially
we might rewrite
  `DEREFERENCE(expr, 0)`
to
  `DEREFERENCE(expr, expr1)`
    `expr1 = 0`

This was convenient since we didn't need to handle constant differently.
However, in RowExpressionInterpreter, it is assumed that the second parameter
of DEREFERENCE will be a number, which is no longer true with the change
in PlanRemoteProjection.

This PR changes the behavior of PlanRemoteProjections to not generate additional
projection for constant.