Avoid unnecessary variable for ConstantExpression in PlanRemoteProjections
The remote project operator expect all inputs of remote functions to be
input references, so in PlanRemoteProjection, we always create variables for
expressions, including constants. However, this would cause plan failure in
case of dereference. Essentially we might rewrite
  `DEREFERENCE(remote_foo(1), 0)`
to
```
  [local] DEREFERENCE(remote_foo, expr1)
    [remote] remote_foo = remote_foo(expr2), expr1 = expr1
      [local] expr1 = 0, expr2 = 1
```

This was convenient since we didn't need to handle constant differently.
However, in RowExpressionInterpreter, it is assumed that the second parameter
of DEREFERENCE will be a number, which is no longer true with the change
in PlanRemoteProjection.

We fix this issue by only adding variables for constants that are inputs for
remote functions, so the above expression will be rewritten into
```
  [local] DEREFERENCE(remote_foo, 0)
    [remote] remote_foo = remote_foo(expr2)
      [local] expr2 = 1
```