Add support to introduce custom prerequisite waiting logic
Implement mechanism to allow for custom logic to be plugged in to make queries to wait
till the time all of the needed prerequisites for a query have been satisfied.

Admins can decide to plugin their custom prerequisite logic by implementing the
`QueryPrerequisites` interface and injecting that by implementing the factory interface
`QueryPrerequisitesFactory` and exposing it through a `Plugin`. If this is not needed,
no action is required since a default no-op implementation will be loaded.