Support parameters in lambda expressions for prepare statement
Override visitLambdaExpression in ParameterExtractingVisitor to support
extracting parameters form lambda expressions. This commit does not add the
override function to DefaultTraversalVisitor because there are other
visitors which do not expect to visit lambdas. For example,
VariablesExtractor doesn't want to extract variables from lambdas and
treat it as dependencies from the source plan.