Remove config hive.s3.use-instance-credentials
This flag was added before IAM role support and is incompatible with role support, as it means "always use instance credentials". After the introduction of DefaultAWSCredentialsProviderChain, instance credentials can still be used even if this flag is false, as they are the last item in the default chain. Given the confusing nature of this flag, and the incompatibilty with existing role support and the upcoming security mapping feature, it seems best to remove the flag entirely.