Compact deserialized page in SingleStreamSpiller
Deserialized pages may contain block that is not
compact (e.g.,VariableWidthBlock might contain
a slice that is a view of all columns), which
results in incorrect retained size and memory
accounting issue.

When a block holds a slice of the entire page,
depends on number of affected blocks in a page,
memory could be off by N times (N >= count of
VariableWidthBlock).