Use segmented Slice in SliceDictionaryWriter
This is a redo of the commit, 53ebcda217e2c3de8b77b8b3c7159e22dd28f7cb

This commit was reverted as caution, but this commit did not introduce
regression. Store elements of dictionary in Segmented Slices, instead of
one contiguous segment. When the number of elements
in the dictionary is less than 100,000 there is no noticeable
performance degradation. When the number of elements in the
dictionary reaches 10,000,000 sorting/comparing the element
needs to compute segment/offset which makes it worse by 10%.
But this is an unlikely case