Add counters for the number of skipped stripes and total stripes
The StripeReader currently always opens all row group protos. We observe
>10% CPU spent in deserializing the protos for some prod workload. One
potential optimization is to make the row group loading lazy and only
open them for the filtered columns if the entire stripe is skipped.

Adding these counters to estimate the potential impact of this
optimization to help decide whether it is worth implementing it.