Choose join side with small tables as build side
This commit adds a new logic to support Cost-Based Optimization for
join distribution type decision when unknown statistics are provided.
If CBO is unable to determine join distribution type due to unknow
stats, try to collect the source table size and broadcast the samller
table. However if there is any JoinNode or Unnest, it will not be
possible to determine the source, optimizer will repartition both side.

Cherry-pick of https://github.com/trinodb/trino/pull/5141

Co-authored-by: Karol Sobczak <napewnotrafi@gmail.com>