Carry Forward Direct Encoding Decision for Stripes
Dictionary encoding is CPU (to maintain dictionary) and memory (size of
dictionary in memory) intensive. Some times, dictionary encoding does
not result in any storage savings and is wasteful. For each stripe
before this change, it will always start with dictionary and later
based on some criteria it will fall back to Direct.

But for some long dictionary (DWRF) use cases this is resulting in
significant CPU and memory regression and OutOfMemory failures. This
change introducees an option to preserve direct encoding decisions
for configurable number of stripes. For those stripes, it will use
direct encoding without attempting dictionary encoding at all. This
is a feature of existing DWRF library implementation and it does not
result in any storage regression for the use cases.