Expose MapBlockBuilder KeyBlockBuilder and ValueBlockBuilder
When Presto writer is used as a library outside of presto, building map
requires writing key and values alternatively. This makes it hard to
vectorize and there is a virtual dispatch per map key and value. On
Nested Map (Map with value as another map, this is 300% slower than
the existing implementations). This change exposes the keyBlockBuilder
and valueBlockBuilder so that caller can write vectorized code (tight
for loops). The new implementation is only marginally better for
existing Presto use cases, so current presto code does not need
migration to this API.