Spill large blocks to separate spill files in distinct aggregates
Distinct Aggregates spilling logic compacts all values corresponding
to the same groupId into a single array block and put it in the output
block of the aggregator. This is done for each aggregate function
in the HashAggregationBuilder. If multiple distinct aggregate functions
are present, then we perform same kind of compaction into
corresponding blocks in the same position within a page.

Due to this, a single row within the page can become really huge and
when this page is spilled, it can fail with 'integer overflow' error during
serialization. Instead of compressing all values within a single array
block, we can spill them into a separate spill file, if the block size goes
beyond a certain threshold.

When large block spill is enabled, we use a hybrid approach for prepping
page for spill. If an array block for a groupId goes beyond a threshold, we
spill it into a separate spill file and store the serialized file handle in the
output block. If the size is within the threshold, we will directly store the
values in the output block. This approach enables us to perform a spill
only iff a block is huge. Since, this will happen rarely, most of the time,
we will be storing the data directly in the output block of the aggregate.