Optimize putIfAbsent in Dictionary Buillders
getHashPositionOfElement is only used by the putIfAbsent. This change
removes the abstraction and some unnecessary lookups. On successful
match, two additional lookups into the hash table is removed. When the
hash position is empty, one additional lookup is removed. Performance
benchmark shows about 15% improvement on the LongDictionary use case.
The SliceDictionary performance improvement is negligible as the String
dictionary is costly to maintain.