Register alluxio cache read metrics
Some of the local cache read metrics like bytes read, hit rate etc are
registered only when a LocalCacheFileInStream object is created as
part of openFile method below. In Presto, when the coordinator is not
acting as a worker, file read might only happen on workers, while
metrics are only exposed from coordinators, resulting in the metrics
not being registered. Always register the metrics when initializing
this class.A fix https://github.com/Alluxio/alluxio/pull/14912 for this
in alluxio project is also in progress. Once we upgrade to alluxio
version with this fix, we can revert this fix.