Use Factory method in OrcWriterOptions
while working on CompressionBufferPool realized that OrcWriterOptions
took in the implementation. presto-raptor for example shared the
same OrcWriterOptions for multiple writer. This required the
implementation to be thread safe or stateless. CompressionBufferPool
was neither and it introduced subtle race condition bugs. To avoid
future surprises moved the OrcWriterOptions to provide a factory.
OrcWriter will create an implementation and each OrcWriter will only
work with one implementation.