Change rehashing strategy in MultiChannelGroupByHash
Previously the hash table capacity was checked every row to see whether a rehash
is needed. Now the input page is split into batches and it is assumed that every
row in batch will create a new group (which is rarely the case) and rehashing
is done in advance before processing.
This may slightly increase memory footprint for small number of groups, however
there is a tiny performance gain as the capacity is not checked every row.

Cherry-pick of https://github.com/trinodb/trino/pull/12336/commits/88cd4920b5a9a0491680fda7410b324e0274da37

Co-authored-by: skrzypo987<krzysztof.skrzypczynski@starburstdata.com>