[native] Fix arrow-flight to be used for IBM internal use
In OSS the name is not branded but internally we use ibm-arrow-flight
instead of arrow-flight because on the Java side the arrow flight
plugin is also called ibm-arrow-flight and keeping a consistent
name in the catalog configs for the connector helps simplify the
configuration.

We also need to link into the custom IBM authenticators.

Alchemy-item: (ID = 900) [native] Fix arrow-flight to be used for IBM internal use commit 1/1 - aae8557128dc5340d0699d5015443bdfef6f12a6