[native] Enable CPR for build and fix Arrow build testing
The IBM SaasAuthenticator requirtes CPR which is only built
when testing is enabled. However, in pipelines
testing is not enabled and the build fails as a result.

For Arrow to be used or the Flight Connector the
ARROW_TESTING library is required. This is enabled
using the BENCHMARK option.

Alchemy-item: (ID = 1159) [native] Enable CPR for build without testing enabled commit 1/1 - 6d9c9f162d88ef47fe50e8072588e806f69b62ca